requirements.txt
beautifulsoup4==4.12.3
click==8.1.7 
python-slugify==8.0.4 
requests==2.32.3